// react-frontend/src/api/models/allergy-entity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AllergyEntity
 */
export interface AllergyEntity {
    /**
     * 
     * @type {number}
     * @memberof AllergyEntity
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof AllergyEntity
     */
    'allergy'?: string;
    /**
     * 
     * @type {string}
     * @memberof AllergyEntity
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof AllergyEntity
     */
    'updatedAt'?: string;
}